+++ src/maneki/components/ManekiLoadingPaper.tsx
