+++ src/maneki/modules/airdrop/AirdropContainer.tsx
@@ -8,8 +8,8 @@ import { ReactElement } from 'react-markdown/lib/react-markdown';
 import { ConnectWalletPaper } from '../../../components/ConnectWalletPaper';
 import { useWeb3Context } from '../../../libs/hooks/useWeb3Context';
 import { marketsData } from '../../../ui-config/marketsConfig';
+import ManekiLoadingPaper from '../../components/ManekiLoadingPaper';
 import { useAirdropContext } from '../../hooks/airdrop-data-provider/AirdropDataProvider';
-import ManekiLoadingPaper from '../../utils/ManekiLoadingPaper';
 import AirdropContentWrapper from './AirdropContentWrapper';
 import AirdropTable from './AirdropTable';
 import randomAddrs from './devRandAddr';
