+++ src/maneki/utils/stringConverter.tsx
